# swagger-spec.json
{"openapi":"3.0.0","info":{"title":"ToDoList","description":"The ToDoList API description","version":"1.0","contact":{}},"tags":[{"name":"User","description":""}],"servers":[],"components":{"schemas":{"User":{"type":"object","properties":{"u_id":{"type":"number","description":"User ID"},"email":{"type":"string","description":"Email"},"password":{"type":"string","description":"Password"},"avatar":{"type":"string","description":"Avatar"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"},"username":{"type":"string","description":"username"}},"required":["email","password","username"]},"userResponseDTO":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"},"token":{"type":"string","description":"JWT token"}},"required":["user","token"]},"LoginDTO":{"type":"object","properties":{"email":{"type":"string","description":"Email","example":"user1@gmail.com"},"password":{"type":"string","description":"Password","example":"Aa123456"}},"required":["email","password"]},"Task":{"type":"object","properties":{"t_id":{"type":"number","description":"Task ID"},"u_id":{"type":"number","description":"user id"},"name":{"type":"string","description":"task name"},"isImportant":{"type":"boolean","description":"this task whether important"},"date":{"format":"date-time","type":"string","description":"task date"},"isFinished":{"type":"boolean","description":"task status"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"}},"required":["u_id","name","date"]}}},"paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/user/register":{"post":{"operationId":"UsersController_register","parameters":[],"requestBody":{"required":true,"description":"the u_id is not required","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/userResponseDTO"}}}}},"tags":["User"]}},"/user/login":{"post":{"operationId":"UsersController_login","parameters":[],"requestBody":{"required":true,"description":"User login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"201":{"description":"success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/userResponseDTO"}}}}},"tags":["User"]}},"/user/update":{"put":{"operationId":"UsersController_update","parameters":[{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"description":"the u_id is required","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"success"}},"tags":["User"]}},"/user/checkEmail":{"get":{"operationId":"UsersController_checkEmail","parameters":[{"name":"email","required":true,"in":"query","description":"email","schema":{"type":"string"}}],"responses":{"200":{"description":"Not exist"},"203":{"description":"email exist"}},"tags":["User"]}},"/user/checkUsername":{"get":{"operationId":"UsersController_checkUsername","parameters":[{"name":"username","required":true,"in":"query","description":"username","schema":{"type":"string"}}],"responses":{"200":{"description":"Not exist"},"203":{"description":"email exist"}},"tags":["User"]}},"/user/findOneByUid":{"get":{"operationId":"UsersController_findOneByUid","summary":"for Taka, data customized","parameters":[{"name":"u_id","required":true,"in":"query","description":"u_id","schema":{"type":"string"}}],"responses":{"200":{"description":"user exist"},"203":{"description":"user not exist"}},"tags":["User"]}},"/task/save":{"post":{"operationId":"TaskController_save","parameters":[],"requestBody":{"required":true,"description":"task information","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"responses":{"201":{"description":"successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}},"tags":["Task"]}},"/task/update":{"put":{"operationId":"TaskController_update","parameters":[],"requestBody":{"required":true,"description":"task information","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"responses":{"201":{"description":"successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}},"tags":["Task"]}},"/task/delete":{"delete":{"operationId":"TaskController_delete","parameters":[{"name":"t_id","required":true,"in":"query","description":"task information delete","schema":{"type":"string"}}],"responses":{"200":{"description":"successfully delete task"}},"tags":["Task"]}},"/task/getAllTaskByUid":{"get":{"operationId":"TaskController_getAllTaskByUid","parameters":[{"name":"u_id","required":true,"in":"query","description":"user id","schema":{"type":"string"}}],"responses":{"200":{"description":"return all tasks belong to this user"}},"tags":["Task"]}},"/task/getAllImportantTaskByUid":{"get":{"operationId":"TaskController_getAllImportantTaskByUid","parameters":[{"name":"u_id","required":true,"in":"query","description":"user id","schema":{"type":"string"}}],"responses":{"200":{"description":"return all important tasks belong to this user"}},"tags":["Task"]}},"/task/getTodayTaskByUid":{"get":{"operationId":"TaskController_getTodayTaskByUid","parameters":[{"name":"u_id","required":true,"in":"query","description":"user id","schema":{"type":"string"}}],"responses":{"200":{"description":"return all today's tasks belong to this user"}},"tags":["Task"]}},"/task/getTasks":{"get":{"operationId":"TaskController_getTasks","parameters":[{"name":"t_id","required":false,"in":"query","description":"Task ID","schema":{"type":"number"}},{"name":"u_id","required":true,"in":"query","description":"user id","schema":{"type":"number"}},{"name":"name","required":true,"in":"query","description":"task name","schema":{"type":"string"}},{"name":"isImportant","required":false,"in":"query","description":"this task whether important","schema":{"type":"boolean"}},{"name":"date","required":true,"in":"query","description":"task date","schema":{"format":"date-time","type":"string"}},{"name":"isFinished","required":false,"in":"query","description":"task status","schema":{"type":"boolean"}},{"name":"created","required":false,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"updated","required":false,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"task","required":true,"in":"query","description":"task condition","schema":{}}],"responses":{"200":{"description":"return all today's tasks belong to this user by specific condition"}},"tags":["Task"]}}}}